Move `CryptoKeySecurityAccessRules` to its own file
